pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turkishvoice"
version = "0.2.0"
description = "Açık kaynak Türkçe Metin-Şarkı Sentezi (TTS) - Gerçek Zamanlı Neural TTS"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "TurkishVoice Team", email = "contact@turkishvoice.ai"}
]
keywords = ["tts", "text-to-speech", "turkish", "turkce", "speech-synthesis", "opentts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "soundfile>=0.12.0",
    "click>=8.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.65.0",
    "edge-tts>=6.1.0",
    "piper-tts>=1.2.0",
    "onnxruntime>=1.15.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
]
webui = [
    "gradio>=3.40.0",
]
clone = [
    "coqui-tts>=0.22.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
all = [
    "turkishvoice[api,webui,clone,dev]",
]

[project.scripts]
turkishvoice = "turkishvoice.cli.main:cli"

[project.urls]
Homepage = "https://github.com/turkishvoice/turkishvoice"
Documentation = "https://turkishvoice.ai/docs"
Repository = "https://github.com/turkishvoice/turkishvoice"
Issues = "https://github.com/turkishvoice/turkishvoice/issues"
Changelog = "https://github.com/turkishvoice/turkishvoice/releases"

[tool.setuptools.packages.find]
where = ["."]
include = ["turkishvoice*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
